Added lib/mini_tools
In order to make the library installable via RubyGems as per a usual
gem, it is necessary to create a single file that requires the
individual modules. Otherwise it would be necessary to require the
modules individually and `require 'mini_tools'` would raise an error
which is not standard behaviour for a gem.